changed around units and time

+++ src/main/java/net/stzups/tanks/game/Bullet.java
@@ -8,7 +8,7 @@ class Bullet extends Entity {
     private static final int MAX_HITS = 3;
     Player owner;
     int hits;
-    Bullet(UUID id, float x, float y, float speed, float direction, float rotation, int width, int height, Player owner, int hits) {
+    Bullet(UUID id, float x, float y, float speed, float direction, float rotation, float width, float height, Player owner, int hits) {
         super(id, x, y, speed, direction, rotation, width, height);
         this.owner = owner;
         this.hits = hits;
